scripts: Use gsed instead of sed
This is needed to make the build work on mac (where BSD sed is instead
used). Is there a way to specify a different binary for sed within the
build system?

PS: It's the first time I send patches and I'm not an expert of
kernel/embedded development, please excuse me if I'm missing some
important netiquette... I read the README and the doc on the Linux
kernel page for patches...